Moved automatic tempdir cleanup to config.py to keep chefs.py readable
