+++ ricecooker/config.py
@@ -1,10 +1,15 @@
-# Settings for rice cooker
+"""
+Settings and global config values for ricecooker.
+"""
+import atexit
 import hashlib
 import logging.config
 import os
-
 import requests
 from requests_file import FileAdapter
+import shutil
+import tempfile
+
 
 UPDATE = False
 COMPRESS = False
@@ -198,6 +203,31 @@ SUSHI_BAR_STAGES_URL = "{domain}/api/channelruns/{run_id}/stages/"
 SUSHI_BAR_PROGRESS_URL = "{domain}/api/channelruns/{run_id}/progress/"
 SUSHI_BAR_LOGS_URL = "{domain}/logs/{run_id}/"
 SUSHI_BAR_CONTROL_URL = "{domain}/control/{channel_id}/"
+
+
+# Automatic temporary direcotry cleanup
+chef_temp_dir = os.path.join(os.getcwd(), '.ricecooker-temp')
+
+@atexit.register
+def delete_temp_dir():
+    if os.path.exists(chef_temp_dir):
+        LOGGER.info("Deleting chef temp files at {}".format(chef_temp_dir))
+        shutil.rmtree(chef_temp_dir)
+
+# While in most cases a chef run will clean up after itself, make sure that if it didn't,
+# temp files from the old run are deleted so that they do not accumulate.
+delete_temp_dir()
+
+# If tempdir is set already, that means the user has explicitly chosen a location for temp storage
+if not tempfile.tempdir:
+    os.makedirs(chef_temp_dir)
+    LOGGER.info("Setting chef temp dir to {}".format(chef_temp_dir))
+    # Store all chef temp files in one dir to avoid issues with temp or even primary storage filling up
+    # because of failure by the chef to clean up temp files manually.
+    tempfile.tempdir = chef_temp_dir
+
+
+
 
 
 # Character limits based on Kolibri models
